requirements.txt
fastapi==0.111.0
numpy==1.21.2
Pillow==7.0.0
torch==2.3.0
torchvision==0.18.0
opencv-contrib-python==4.5.3.56
requests==2.22.0
pytest==8.2.1